Getting rid of compiler warnings on PC

+++ Source/Copier.cpp
@@ -33,7 +33,7 @@ Copier::~Copier()
 
 void Copier::buttonClicked(Button* b)
 {
-	int multiplier = pow(2, buttons.indexOf( b ));
+	int multiplier = int(pow(2, buttons.indexOf( b )));
 	//let the listeners know
 	Component::BailOutChecker checker (this);
 	if (! checker.shouldBailOut())
@@ -49,9 +49,9 @@ void Copier::paint (Graphics& g)
 
 void Copier::resized()
 {
-	float h = 1.0 / buttons.size() - 0.01;
+	float h = 1.0f / buttons.size() - 0.01f;
 	for ( int i = 0; i < buttons.size(); i++ )
 	{
-		buttons[i]->setBoundsRelative(0.01, 0.01 + i * h + 0.01, 0.98, h);
+		buttons[i]->setBoundsRelative(0.01f, 0.01f + i * h + 0.01f, 0.98f, h);
 	}
 }
